+++ src/nvim/sign.h
@@ -1,0 +1,10 @@
+#ifndef NVIM_SIGN_H
+#define NVIM_SIGN_H
+
+#include <stdbool.h>
+
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "sign.h.generated.h"
+#endif
+#endif  // NVIM_SIGN_H
